move last checkpoint hint
