+++ kernel/src/lib.rs
@@ -130,6 +130,11 @@ pub mod log_segment;
 #[cfg(not(feature = "internal-api"))]
 pub(crate) mod log_segment;
 
+#[cfg(feature = "internal-api")]
+pub mod last_checkpoint_hint;
+#[cfg(not(feature = "internal-api"))]
+pub(crate) mod last_checkpoint_hint;
+
 pub(crate) mod listed_log_files;
 
 #[cfg(feature = "internal-api")]
